lib done and little things handled

+++ ft_memchr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memchr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cprojean <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: cprojean <cprojean@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 11:54:32 by cprojean          #+#    #+#             */
-/*   Updated: 2022/11/28 16:41:33 by cprojean         ###   ########.fr       */
+/*   Updated: 2022/11/28 17:36:20 by cprojean         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*array;
 	unsigned char	cast;
 
 	cast = (unsigned char) c;
 	array = (unsigned char *) s;
 	i = 0;
-	while ((size_t) i < n)
+	while (i < n)
 	{
 		if (array[i] == cast)
 			return (&array[i]);
