+++ ft_memset.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memset.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cprojean <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: cprojean <cprojean@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 12:55:44 by cprojean          #+#    #+#             */
-/*   Updated: 2022/11/28 16:41:50 by cprojean         ###   ########.fr       */
+/*   Updated: 2022/11/28 17:34:39 by cprojean         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int		index;
+	size_t	index;
 	char	*str;
 
 	str = b;
 	index = 0;
-	while ((size_t)index < len)
+	while (index < len)
 	{
 		str[index] = c;
 		index++;
